E-commerce_AI_Agent/app/requirements.txt
# requirements.txt

# FastAPI server
fastapi==0.110.0
uvicorn==0.29.0

# Google Gemini API client
google-generativeai==0.5.2

# PostgreSQL connector
psycopg2-binary==2.9.9

# For environment variables
python-dotenv==1.0.1

# Optional (for enhanced error logging/debugging)
httpx==0.27.0
